% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization_repos.R
\name{org_repos}
\alias{org_repos}
\title{Get name, link and language for all the repositories in the organization.}
\usage{
org_repos(organization, auth = TRUE, gtoken = NULL)
}
\arguments{
\item{organization}{of interest on GitHub}

\item{auth}{optional argument if authentication needed or not}

\item{gtoken}{optional argument - user authentication done using gh_auth.}
}
\value{
the names link and language for all the repositories in the organization
}
\description{
The function returns the name, link and language used for all the repositories in the organization.
}
\examples{
organization_members(organization = "UBC-MDS", auth = TRUE, gtoken = token)

}
