% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organization_members.R
\name{organization_members}
\alias{organization_members}
\title{Get programming languages used/forked by a user on Github.}
\usage{
organization_members(organization, auth = FALSE, gtoken = NULL)
}
\arguments{
\item{organization}{of the interest}

\item{auth}{optional argument if authentication needed or not}

\item{gtoken}{optional argument - user authentication done using gh_auth.}
}
\value{
Returns a list of all the public users with their most commonly used languages in a dataframe and as a ggplot item.
}
\description{
The function returns a list of all the public user in an Organization
}
\examples{
organization_members("UBC-MDS", TRUE)

}
