% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_count.R
\name{repo_count}
\alias{repo_count}
\title{Get the number of Pulic Repositories for an ID}
\usage{
repo_count(user, auth = FALSE, gtoken = NULL)
}
\arguments{
\item{user}{GitHub user ID for the target user.}

\item{auth}{optional argument if authentication needed or not}

\item{gtoken}{optional argument - user authentication done using gh_auth.}
}
\value{
Returns a number of Public Repos for an ID.
}
\description{
The function returns the number of repositories for the given user.
}
\examples{
organization_members("UBC-MDS", TRUE)

}
