% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/user_preferences.R
\name{user_preferences}
\alias{user_preferences}
\title{Get programming languages used/forked by a user on Github.}
\usage{
user_preferences(id, gtoken = NULL)
}
\arguments{
\item{id}{username of the GitHub user}

\item{gtoken}{optional argument - user authentication done using gh_auth.}
}
\value{
a list of: dataframe and ggplot item for the languages used by the user.
}
\description{
Returns the user's language preferences
}
\examples{
user_preferences('sarora','') or user_preferences('sarora')

}
